+++ src/application/tests/testResponseBodyContent.test.ts
@@ -54,6 +54,18 @@ describe('testResponseBodyContent', () => {
     expect(pmTest.script.exec).toMatchSnapshot()
   })
 
+  it('should add content test for property check & null value', async () => {
+    const contentTests = [
+      {
+        key: 'data[0].company_name',
+        value: null
+      }
+    ]
+    pmOperation = testResponseBodyContent(contentTests, pmOperation)
+    const pmTest = pmOperation.getTests()
+    expect(pmTest.script.exec).toMatchSnapshot()
+  })
+
   it('should add content test for property check & empty value', async () => {
     const contentTests = [
       {