+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- ContentTests - Added support for null values for response body (#414)
+
 ## v1.18.1 - (2022-09-30)
 
 - Fix unwanted schema validation (#392)